InputDataType: move to own folder
in order to be able to add more files in the future which are strongly related
