+++ src/eu_unemployment_prediction/input_data_type/__init__.py
@@ -1,0 +1,4 @@
+from ._input_data_type_definition import InputDataTypeDefinition
+from ._input_data_type import InputDataType
+
+__all__ = ["InputDataType", "InputDataTypeDefinition"]
