+++ test/eu_unemployment_prediction/input_data_type/test_input_data_type_data_loading.py
